backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
pillow>=10.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
